sat-benchmark-suite/requirements.txt
# SAT Benchmark Suite - Requirements
# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlite3  # Built-in

# Plotting and Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Scientific Computing
scipy>=1.11.0

# Progress and Monitoring
tqdm==4.66.1

# File Handling
psutil==5.9.6
py7zr==0.20.8
rarfile==4.1

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil==2.8.2
humanize==4.9.0

# Report Generation
reportlab==4.0.7
markdown==3.5.1
